+++ package.json
@@ -52,6 +52,7 @@
     "react-navigation-backhandler": "^2.0.0",
     "react-redux": "^7.2.0",
     "redux": "^4.0.5",
+    "redux-devtools-extension": "^2.13.8",
     "redux-thunk": "^2.3.0",
     "rn-fetch-blob": "^0.12.0",
     "uuid": "^3.3.3",
