+++ service.js
@@ -1,0 +1,4 @@
+module.exports = async function() {
+    // This service needs to be registered for the module to work
+	// but it will be used later in the "Receiving Events" section
+}
