+++ src/redux/reducer.js
@@ -31,13 +31,14 @@ import {
   LOAD_IMAGES,
   LOAD_IMAGES_ROLLBACK,
   LOAD_IMAGES_COMMIT,
+  SET_VALIDATE,
 } from './actions';
 
 const defaultLanguage = getDeviceLanguage();
 //Login
 
 export const user = (
-  state = {token: null, status: null, username: null, role: null},
+  state = {token: null, status: null, username: null, role: null, interactive_help:null},
   action,
 ) => {
   switch (action.type) {
@@ -54,7 +55,13 @@ export const user = (
         token: null,
         username: null,
         role: null,
+        interactive_help: null
       };
+    case SET_VALIDATE:
+      return {
+          ...state,
+          interactive_help:action.interactive_help
+        };
     default:
       return state;
   }
