+++ src/screens/lifemap/Question.js
@@ -16,6 +16,7 @@ import globalStyles from '../../globalStyles';
 
 import TrackPlayer from 'react-native-track-player';
 import RNFetchBlob from 'rn-fetch-blob'
+import Audio from '../../components/Audio';
 
 let dirs = RNFetchBlob.fs.dirs
 
@@ -33,19 +34,11 @@ export class Question extends Component {
 
   state = {
     showDefinition: false,
-    isPlaying: false,
-    currentTrackId: null,
-    donePlaying: false
   };
 
-  onQueueEnd = null;
-  track = {};
 
  
 
-  getProperSourceForOS(source) {
-    return Platform.OS === 'android' ? 'file://' + source : '' + source
-  }
 
 
   getDraft = () =>
@@ -191,45 +184,6 @@ export class Question extends Component {
 
   async componentDidMount() {
     const draft = this.getDraft();
-    console.log('ComponentDidMount Question')
-    console.log(this.indicator)
-    this.track = {
-      url: this.getProperSourceForOS(
-        `${dirs.DocumentDir}/${this.indicator.questionAudio.replace(/https?:\/\//, '')}`
-      ) ,
-      id: this.indicator.id
-    };
-    TrackPlayer.destroy();
-    TrackPlayer.setupPlayer();
-    const trackPlayerCapabilities = [
-      TrackPlayer.CAPABILITY_PLAY,
-      TrackPlayer.CAPABILITY_PAUSE,
-      TrackPlayer.CAPABILITY_SEEK_TO,
-      TrackPlayer.CAPABILITY_JUMP_BACKWARD,
-      TrackPlayer.CAPABILITY_JUMP_FORWARD,
-    ];
-    TrackPlayer.updateOptions({
-      stopWithApp: true,
-      capabilities: trackPlayerCapabilities,
-      notificationCapabilities: [...trackPlayerCapabilities, TrackPlayer.CAPABILITY_SEEK_TO],
-      jumpInterval: 10
-    });
-    // TrackPlayer.registerPlaybackService(() => require('./service.js'));
-    await TrackPlayer.add(this.track);
-    const current = await TrackPlayer.getCurrentTrack();
-    console.log('current', current)
-
-    this.setState({ isPlaying: false, currentTrackId: current });
-
-    this.onQueueEnd = TrackPlayer.addEventListener('playback-queue-ended', async (data) => {
-      console.log('Queue ended');
-      this.setState({ isPlaying: false, isPlaying: false, donePlaying: true });
-      await TrackPlayer.stop();
-
-    })
-
-
-
 
     this.props.updateDraft({
       ...draft,
@@ -245,38 +199,20 @@ export class Question extends Component {
     });
   }
 
-  async togglePlayPause() {
-    if (this.state.isPlaying) {
-      await TrackPlayer.pause();
-      this.setState({ isPlaying: false });
 
-    } else {
-      if (this.state.donePlaying) {
-        await TrackPlayer.skip(this.state.currentTrackId);
-        this.setState({ isPlaying: false, donePlaying: false });
-      }
-      await TrackPlayer.play();
-      this.setState({ isPlaying: true });
-      this.setState({ donePlaying: false });
-    }
-  }
 
   shouldComponentUpdate() {
     return this.props.navigation.isFocused();
   }
 
-  componentWillUnmount() {
-    TrackPlayer.stop();
-    TrackPlayer.reset();
-    TrackPlayer.destroy();
-  }
+  
 
   render() {
     const draft = this.getDraft();
     // added a popup component to the Question.js instead of adding it to the
     // modals folder because it is really smol and does not do much
 
-    const { t } = this.props;
+    const { t, user } = this.props;
 
     return (
       <View style={{ flex: 1 }}>
@@ -341,27 +277,10 @@ export class Question extends Component {
               />
             ) : null}
 
-            {this.state.isPlaying ?
-              <Icon id="player" name="pause-circle-filled" onPress={() => { 
-                console.log('stop')
-                this.togglePlayPause();
-              }} style={{
-                color: colors.palegreen,
-                position: 'absolute',
-                top: '55%',
-                left: '20%',
-              }} size={40} />
-              :
-              <Icon id="player" name="play-circle-filled" onPress={() => {
-                console.log('play')
-                this.togglePlayPause();
-              }} style={{
-                color: colors.palegreen,
-                position: 'absolute',
-                top: '55%',
-                left: '20%',
-              }} size={40} />
+            {user.interative_help &&  this.indicator && this.indicator.questionAudio &&
+              <Audio audioId ={this.indicator.id} url={this.indicator.questionAudio}  />
             }
+            
 
 
             {this.indicator.required ? (
@@ -407,9 +326,10 @@ Question.propTypes = {
   updateDraft: PropTypes.func.isRequired,
 };
 
-const mapStateToProps = ({ dimensions, drafts }) => ({
+const mapStateToProps = ({ dimensions, drafts, user }) => ({
   dimensions,
   drafts,
+  user
 });
 
 const mapDispatchToProps = { updateDraft };
